backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
yt-dlp
openai-whisper
google-generativeai
python-dotenv
websockets
